Organize helper functions into sorted files
